+++ R/eq_location_clean.R
@@ -1,0 +1,13 @@
+#' Clean NOAA data frame
+#'
+#' @param data A raw data frame of NOAA data
+#'
+#' @return a [tibble][tibble::tibble-package]
+#' @export
+#'
+#' @examples
+#' data(noaa)
+#' tidy_noaa <- eq_location_clean(noaa)
+eq_location_clean <- function(data) {
+    data
+}