+++ NEWS.md
@@ -1,4 +1,10 @@
-* raw-data imported from <https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1> (issue #1)
+* Created `eq_location_clean()` function and corresponding tests.
+* Added support for tibbles.
+* NOAA raw-data imported and documented into the package data like a
+  data frame.
+* raw-data imported from
+  <https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1>
+  (issue #1).
 
 # devrcap 0.0.1