+++ NEWS.md
@@ -1,3 +1,4 @@
+* Created `good_date()` to convert date to Date even if negative
 * Created `eq_clean_data()` function and corresponding tests.
 * `eq_location_clean()` function created and tested.
 * Added support for tibbles.