+++ NEWS.md
@@ -1,3 +1,5 @@
+# devrcap 0.1.0
+
 * Removed integration for R 3.1 and 3.2
 * Created `good_date()` to convert date to Date even if negative
 * Created `eq_clean_data()` function and corresponding tests.